Rename SqlCompile to QueryCompile, better naming and file location
